Added markdowns and moved models and classes to a common directory
