+++ evidence_detection/attention_layer.py
