+++ evidence_detection/early_stopping.py
